Rename Microsoft.Devices.Sensors namespace to MonoGame.Framework.Devices.Sensors (#8681)
As per subject line.
