--- MonoGame.Framework/Microsoft.Devices.Sensors.targets
@@ -1,15 +1,0 @@
-<Project>
-  <ItemGroup>
-    <Compile Include="Microsoft\Devices\Sensors\AccelerometerFailedException.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\AccelerometerReading.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\CalibrationEventArgs.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\CompassReading.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\ISensorReading.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\SensorBase.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\SensorFailedException.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\SensorReadingEventArgs.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\SensorState.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\Accelerometer.cs" />
-    <Compile Include="Microsoft\Devices\Sensors\Compass.cs" />
-  </ItemGroup>
-</Project>
