+++ MonoGame.Framework/MonoGame.Framework.Native.csproj
@@ -22,7 +22,7 @@
     <Compile Remove="Utilities\System.Numerics.Vectors\**\*" />
     <Compile Include="Platform\Native\**\*" />
     <Compile Include="Platform\Utilities\ReflectionHelpers.Default.cs" />
-    <Compile Remove="Microsoft\**\*" />
+    <Compile Remove="MonoGame\**\*" />
   </ItemGroup>
 
   <ItemGroup>
@@ -51,5 +51,5 @@
     <Folder Include="Platform\Graphics\Vertices\" />
   </ItemGroup>
 
-  <Import Project="Microsoft.Devices.Sensors.targets" />
+  <Import Project="MonoGame.Framework.Devices.Sensors.targets" />
 </Project>
