+++ MonoGame.Framework/MonoGame/Framework/Devices/Sensors/CalibrationEventArgs.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Devices.Sensors
+namespace MonoGame.Framework.Devices.Sensors
 {
     /// <summary>
     /// Provides data for Calibrate and events.
