+++ MonoGame.Framework/MonoGame/Framework/Devices/Sensors/CompassReading.cs
@@ -5,7 +5,7 @@
 using System;
 using Microsoft.Xna.Framework;
 
-namespace Microsoft.Devices.Sensors
+namespace MonoGame.Framework.Devices.Sensors
 {
     /// <summary>
     /// Provides properties for compass readings.
