+++ MonoGame.Framework/MonoGame/Framework/Devices/Sensors/ISensorReading.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Devices.Sensors
+namespace MonoGame.Framework.Devices.Sensors
 {
     /// <summary>
     /// Common interface used for sensor readings.
