+++ MonoGame.Framework/MonoGame/Framework/Devices/Sensors/SensorFailedException.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Devices.Sensors
+namespace MonoGame.Framework.Devices.Sensors
 {
     /// <summary>
     /// Represents errors related to sensors.
