+++ MonoGame.Framework/MonoGame/Framework/Devices/Sensors/SensorState.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Devices.Sensors
+namespace MonoGame.Framework.Devices.Sensors
 {
     /// <summary>
     /// Defines the state of the sensor.
