+++ MonoGame.Framework/Platform/Android/Devices/Sensors/Accelerometer.cs
@@ -7,7 +7,7 @@ using Android.Content;
 using Android.Hardware;
 using Microsoft.Xna.Framework;
 
-namespace Microsoft.Devices.Sensors
+namespace MonoGame.Framework.Devices.Sensors
 {
     /// <summary>
     /// Provides Android applications access to the device's accelerometer sensor.
