+++ MonoGame.Framework/Platform/Desktop/Devices/Sensors/Compass.cs
@@ -5,7 +5,7 @@
 using System;
 using Microsoft.Xna.Framework;
 
-namespace Microsoft.Devices.Sensors
+namespace MonoGame.Framework.Devices.Sensors
 {
     /// <summary>
     /// Provides Desktop applications access to the device's compass sensor.
